# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: TOEM HERE API
  version: 0.2.0
servers:
  - url: http://localhost:8080
paths:
...
  /v1/auth/bootstrap:
    post:
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/User' }
        '401': { $ref: '#/components/responses/Error' }
  /v1/users/me:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/User' }
        '401': { $ref: '#/components/responses/Error' }
    patch:
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                username: { type: string, pattern: '^[A-Za-z0-9_]{3,30}$' }
                displayName: { type: string, minLength: 1, maxLength: 80 }
                locale: { type: string, pattern: '^[a-z]{2}(-[A-Z]{2})?$' }
      responses:
        '200': { $ref: '#/components/responses/User' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        AUTH_MODE=supabase: a Supabase-issued RS256 JWT verified through JWKS.
        AUTH_MODE=local (development/test only, never production): the fixed
        local development credential. The server derives the user in both modes.
  schemas:
    User:
      type: object
      required: [id, username, displayName, avatarUrl, locale, roles, profileComplete]
      properties:
        id: { type: string, format: uuid }
        username: { type: [string, 'null'] }
        displayName: { type: [string, 'null'] }
        avatarUrl: { type: [string, 'null'], format: uri }
        locale: { type: string }
        roles: { type: array, items: { type: string } }
        profileComplete: { type: boolean }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
...
  responses:
...
    User:
      description: Current internal TOEM HERE user
      content: { application/json: { schema: { $ref: '#/components/schemas/User' } } }
    Error:
      description: Consistent API error
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }